+++ src/app/shared/RelayTitle.svelte
@@ -14,7 +14,7 @@
   let showStatus = false
 
   onMount(() => {
-    return poll(10_000, () => {
+    return poll(3000, () => {
       const socket = Network.pool.get(relay.url, {autoConnect: false})
 
       meta = socket?.meta